[REFACTOR] FirebaseTokenManager -> DeviceTokenManagerUseCase
